requirements.txt
# Core dependencies
qbittorrent-api==2025.2.0
PyYAML==6.0

# Runtime requirements
python>=3.8, <4.0
